Modified metadata port_id to uint16 for loopback mode

+++ pkg/framework/dataplane/loopback.go
@@ -121,7 +121,7 @@ func convertToPktOut(port uint32, pkt []byte) *v1.PacketOut {
 	po.Payload = pkt
 	//MetadataId 1 represents egress_physical_port of packet_out controller_packet_metadata based on the P4 program
 	//TODO: make the metadata id configurable to allow loopback for any P4 program
-	po.Metadata = []*v1.PacketMetadata{{MetadataId: 1, Value: common.GetUint32(port)}}
+	po.Metadata = []*v1.PacketMetadata{{MetadataId: 1, Value: common.GetUint16(port)}}
 	return po
 }
 
@@ -131,7 +131,7 @@ func convertToPktIn(port uint32, pkt []byte) *v1.PacketIn {
 	pi.Metadata = []*v1.PacketMetadata{
 		//MetadataId 1 represents ingress_physical_port of packet_in controller_packet_metadata based on the P4 program
 		//TODO: make the metadata id configurable to allow loopback for any P4 program
-		{MetadataId: 1, Value: common.GetUint32(port)},
+		{MetadataId: 1, Value: common.GetUint16(port)},
 	}
 	return pi
 }
